+++ src/components/MapView.jsx
@@ -5,8 +5,16 @@ import { db } from "../firebase";
 
 import GoogleMapReact from 'google-maps-react-markers';
 
+import ReactMarkdown from 'react-markdown';
+
+import {imageURL} from '../utils/utils';
+
 // MUI
+import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
+import Link from '@mui/material/Link';
+import Button from '@mui/material/Button';
+import Chip from '@mui/material/Chip';
 import IconButton from '@mui/material/IconButton';
 import PushPinOutlinedIcon from '@mui/icons-material/PushPinOutlined';
 import StarOutlinedIcon from '@mui/icons-material/StarOutlined';
@@ -14,9 +22,38 @@ import MenuBookOutlinedIcon from '@mui/icons-material/MenuBookOutlined';
 import LocalCafeOutlinedIcon from '@mui/icons-material/LocalCafeOutlined';
 import LocalLibraryOutlinedIcon from '@mui/icons-material/LocalLibraryOutlined';
 
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogTitle from '@mui/material/DialogTitle';
+
 // Custom UI
 import MapPin from './MapPin';
 import Spinner from './Spinner';
+import EventsDetailsImage from './EventsDetailsImage';
+
+
+// Theme helpers
+import { useTheme } from '@mui/material/styles';
+import useMediaQuery from '@mui/material/useMediaQuery';
+
+const styleEventTitle={
+  textAlign: "center"
+}
+
+const styleFacilities={
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
+  gap: "1rem",
+}
+
+const styleEventDecsription={
+  backgroundColor: "#f5f5f5",
+  padding: "1rem",
+  marginBottom: "1rem",
+  marginTop: "1rem"
+}
 
 const styleMap={
   display: "block",
@@ -46,16 +83,25 @@ const styleFilter = {
 
 export default function MapView() {
 
+  const theme = useTheme();
+  const fullScreen = useMediaQuery(theme.breakpoints.down('sm'));
+
   const [locations, setLocations] = useState([]);
   const [filteredLocations, setFilteredLocations] = useState([]);
   const [isLoaded, setIsLoaded] = useState(false);
 
+  const [isOpenDialog, setIsOpenDialog] = useState(false);
+
   const mapRef = useRef(null)
   const [mapReady, setMapReady] = useState(false)
 
-  const q = query(collection(db, "locations"));
+  const [pinTitle, setPinTitle] = useState('');
+  const [pinFacilities, setPinFacilities] = useState([]);
+  const [pinDescription, setPinDescription] = useState('');
+  const [pinImage, setPinImage] = useState('');
 
   const getLocations = async () => {
+    const q = query(collection(db, "locations"));
     const querySnapshot = await getDocs(q);
     const l = [];
     querySnapshot.forEach((doc) => {
@@ -79,6 +125,24 @@ export default function MapView() {
     getLocations();
   }, [])
 
+  const onClickPin = (data) => {
+    setPinTitle(data.title);
+    setPinDescription(data.description);
+    setPinImage(data.image);
+    if (data.facilities) {
+      setPinFacilities(data.facilities.split(','));
+    }
+    setIsOpenDialog(true);
+  }
+
+  const handleCloseDetails = () => {
+    setPinTitle('');
+    setPinDescription('');
+    setPinImage('');
+    setPinFacilities([]);
+    setIsOpenDialog(false);
+  };
+
   const defaultLocation = {
     center: {
       lat: 55.9579741,
@@ -94,6 +158,36 @@ export default function MapView() {
 
   return (
     <>
+      <Dialog
+        fullWidth
+        maxWidth="sm"
+        fullScreen={fullScreen}
+        open={isOpenDialog}
+        onClose={handleCloseDetails}
+        scroll="paper"
+        aria-labelledby="add-dialog-title">
+        <DialogTitle id="add-dialog-title" sx={styleEventTitle}>
+          <Typography variant="h2" component="span">{pinTitle}</Typography>
+        </DialogTitle>
+        <DialogContent>
+          <Box style={styleEventDecsription}>
+            <ReactMarkdown children={pinDescription} />
+            {/* <Typography component="p" variant='p' sx={{ mt: 2 }}>More at: <Link to={currentEvent.url}>{cleanUrl(currentEvent.url)}</Link></Typography> */}
+          </Box>
+          <Box style={styleFacilities}>
+            { pinFacilities.map((facility, index) => (
+              <Chip key={index} label={facility} />
+            ))}
+          </Box>
+          { pinImage &&
+            <EventsDetailsImage image={imageURL(pinImage?.image, 'medium')} alt={pinTitle} />
+          }
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={handleCloseDetails} variant='contained'>Close</Button>
+        </DialogActions>
+      </Dialog>
+
       { isLoaded &&
         <>
           <Box sx={styleFilter}>
@@ -124,9 +218,11 @@ export default function MapView() {
 
               {filteredLocations.map((place) => (
                 <MapPin
+                  onClickPin={onClickPin}
                   key={place.id}
                   tags={place.tags}
                   title={place.title}
+                  data={place}
                   lat={parseFloat(place.lat)}
                   lng={parseFloat(place.lng)}
                 />